% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce.R
\name{announce}
\alias{announce}
\alias{print.announce}
\title{Announce Class for Consistent Printing}
\usage{
announce(object = vector(), lead = "Announce", ...)

\method{print}{announce}(x, ...)
}
\arguments{
\item{object}{Object to be converted to \code{"announce"} class.}

\item{lead}{a \code{character} string giving the title to be printed.}

\item{\dots}{other named arguments to be forwarded to print methods of classes inherited from \code{object}.}
}
\value{
an object of class \code{"announce"} inheriting from class(es) of \code{object}.
}
\description{
Creates an object of class \code{"announce"} with a built-in title string used for printing.
}
\details{
\code{announce()} converts an object to class \code{"announce"}, inheriting from its existing class(es).
}
\examples{
announce()
(cpt <- announce("x", lead = "Lorum ipsum dolor sit amet"))
.class2(cpt)

## an Announce object, or one inheriting from announce, can be safely overwritten
(cpt <- announce(cpt, "Consectetur adipiscing elit"))
.class2(cpt)

rm(cpt)
}
\seealso{
\code{\link[base:cat]{cat()}}, \code{\link[base:print]{print()}}.

Other print: 
\code{\link{lf}()},
\code{\link{print_all}()}
}
\concept{print}
\keyword{print}
