% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-glm.R
\name{anova_tbl}
\alias{anova_tbl}
\title{Create Tibble from List of Anovas}
\usage{
anova_tbl(anova_ls)
}
\arguments{
\item{anova_ls}{a \code{\link{list}} of \code{anova} objects.}
}
\value{
An object of classes \code{"anova_tbl"}, \code{"announce"}, inheriting from \code{\link[tibble:tibble-package]{tibble}}, showing the
table entries for each anova, one per line.
}
\description{
Create a tibble from a list of anovas that compare a model to the null model or that compare two nested models.
}
\details{
The anovas to be compared must all be of the same type i.e., all must be analyses of a single model object or
all must be comparisons of two models. If any of the models has more than one independent variable, the results
may be difficult to interpret and a warning will be given.

\code{anova_tbl()} can be used to be easily and conveniently compare a list of anovas obtained with
\code{\link{summanov}} and  \code{\link[purrr:list_transpose]{list_transpose}}, see examples.
}
\examples{
## Following on from summanov() examples… 
## Simulate Bernoulli data
(d <- list(
    iv2 = list(g = c("a", "c", "e"), h = c("b", "d", "f")),
    iv3 = list(i = c("a", "b", "c"), j = c("d", "e", "f")),
    iv4 = list(k = c("a", "b"), l = c("c", "d"), m = c("e", "f"))
) |> add_grps(binom_data(levels = 6), iv, .key = _))

## Create list of GLM anovas using summanov()
(alist <- d |> summanov(cbind(pn, qn), starts_with("iv")) |>
    list_transpose() |> _$anova)

## Tibble from list of GLM anovas 
alist |> anova_tbl()

## Add GLM anova with two independent variables
alist$iv3_iv4 <- glm(cbind(pn, qn) ~ iv3 + iv4, family = binomial, data = d) |>
    anova()

## Tibble from list with multivariable anova invokes warning
try(alist |> anova_tbl())

rm(d, alist)
}
\seealso{
\code{\link[stats:anova.glm]{anova.glm()}} \code{\link[purrr:list_transpose]{list_transpose()}}, and
\code{\link[tibble:tibble-package]{tibble}}.

Other comp_glm: 
\code{\link{comp_glm}()},
\code{\link{summanov}()},
\code{\link{univ_anova}()}
}
\concept{comp_glm}
\keyword{models}
\keyword{regression}
