% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency-table.R
\name{binom_contingency}
\alias{binom_contingency}
\title{Binomial Contingency Table for Data with a Binary Outcome}
\usage{
binom_contingency(
  .data,
  .dep_var,
  ...,
  .drop_zero = FALSE,
  .propci = FALSE,
  .level = 0.95
)
}
\arguments{
\item{.data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of a binary dependent variable, which should be
\code{numeric} with values of \var{0} and \var{1}.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> quoted name(s) of one or more \code{factor}s or
\verb{character vector}s in \code{.data}, to be included (or excluded) in the output.}

\item{.drop_zero}{\code{logical}. If \code{TRUE}, \code{\link[base:levels]{levels}} of explanatory factors for which values of
\code{.dep_var} are either all zero or all one are dropped from the output; default \code{FALSE}.}

\item{.propci}{\code{logical}. If \code{TRUE}, each row of the output \code{"binom_contingency"} object includes totals, proportions
and confidence intervals; default \code{FALSE}.}

\item{.level}{the confidence level required; default \var{0.95}.}
}
\value{
An object of class \code{"binom_contingency"}, \code{"announce"}, inheriting from \code{\link[tibble:tibble-package]{tibble}},
with columns \code{pn} and \code{qn} representing the number of "successes" and "failures" respectively, and further columns
for independant (explanatory) variables.

If \code{.propci = TRUE} additional columns are output representing totals, proportions and confidence intervals.
}
\description{
\code{\link[=binom_contingency]{binom_contingency()}} creates a binomial contingency table for data with a binary dependent variable and
one or more categorical independent variables, optionally including totals, proportions and confidence intervals.
}
\details{
Categorical variables (i.e. factors or character vectors) in \code{.data} required as factors in the resulting
contingency table may be selected for inclusion or exclusion using the \code{\dots} argument and the
<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax from package \pkg{\link[dplyr]{dplyr}}, including use of
\dQuote{selection helpers}. If no \code{\dots} arguments are supplied, all categorical variables in \code{.data}
(other than \code{.dep_var}) will be used.

A \link[base:list]{list} of \link[rlang:topic-defuse]{defused R expressions}, as for instance created by
\code{\link[=expl_fcts]{expl_fcts()}}, may be used as the \code{\dots} arguments and should be injected using the
\link[rlang:splice-operator]{splice-operator}, \verb{!!!}, see examples.

Use \code{drop_zero = TRUE} to drop \code{\link[base:levels]{levels}} of explanatory factors for which values of
\code{.dep_var} are either all zero or all one, to prevent a warning messages that ‘fitted probabilities
numerically 0 or 1 occurred’ when fitting generalized linear models using \code{\link[stats:glm]{glm()}} or calculating
odds ratios using \code{\link[=odds_ratio]{odds_ratio()}}; see examples and Venables & Ripley (2002, pp. 197–8).
}
\note{
Confidence intervals are calculated using \code{\link[stats:prop.test]{prop.test()}}, and are based on Wilson's score method
\emph{without} continuity correction (Newcombe, 1998).
}
\examples{
## Bernoulli data with a single explanatory variable
(d <- bernoulli_data())
d |> binom_contingency(dv)
d |> binom_contingency(dv, .propci = TRUE)

## Bernoulli data for a single explanatory variable with levels at which responses are all zero
(d <- bernoulli_data(probs = seq(0.4, 0, length.out = 5)))
d |> binom_contingency(dv)
## Invokes warning: -
##   '! glm.fit: fitted probabilities numerically 0 or 1 occurred'
try(d |> binom_contingency(dv) |>
    odds_ratio(.ind_var = iv))

##  Argument .drop_zero = TRUE in binomial contingency() prevents this warning
d |> binom_contingency(dv, .drop_zero = TRUE)
d |> binom_contingency(dv, .drop_zero = TRUE) |>
    odds_ratio(.ind_var = iv)

## Bernoulli data with multiple explanatory variables
(d <- list(
    iv2 = list(i = c("a", "c", "e", "g"), j = c("b", "d", "f", "h")),
    iv3 = list(k = c("a", "b", "c", "d"), l = c("e", "f", "g", "h")),
    iv4 = list(k = c("a", "b"), l = c("c", "d"), m = c("e", "f"))
) |> add_grps(bernoulli_data(levels = 8), iv, .key = _))

d |> binom_contingency(dv)
d |> binom_contingency(dv, iv, iv3)
d |> binom_contingency(dv, !c(iv2, iv4))
d |> binom_contingency(dv, !!!expl_fcts(d))

d |> binom_contingency(dv, .propci = TRUE)
d |> binom_contingency(dv, .drop_zero = TRUE)

d |>
   binom_contingency(dv, iv2, iv3, .drop_zero = TRUE) |>
   glm(cbind(pn, qn) ~ ., binomial, data = _) |>
   summary()

d |>
   binom_contingency(dv, iv2, iv3, .drop_zero = TRUE) |>
   glm(cbind(pn, qn) ~ ., binomial, data = _) |>
   odds_ratio()

## Use {dplyr} selection helpers e.g., last_col(), num_range() and starts_with()
d |> binom_contingency(dv, last_col(1L))  ## Offset of 1L used, since last column of d is dv
d |> binom_contingency(dv, !last_col())
d |> binom_contingency(dv, num_range("iv", 2:3))
d |> binom_contingency(dv, !num_range("iv", 2:3))
d |> binom_contingency(dv, starts_with("iv"))
d |> binom_contingency(dv, !starts_with("iv")) ## Here, negation excludes all explanatory factors

rm(d)
}
\references{
Confidence interval from R's \code{prop.test()} differs from hand calculation and result from SAS.
\href{https://stats.stackexchange.com/questions/183225/confidence-interval-from-rs-prop-test-differs-from-hand-calculation-and-resul/183238#183238}{
Stack Exchange}.

Newcombe R.G. (1998). Two-Sided Confidence Intervals for the Single Proportion: Comparison of Seven Methods.
\emph{Statistics in Medicine}, \strong{17}, 857-872.
\href{https://doi.org/10.1002/(SICI)1097-0258(19980430)17:8<857::AID-SIM777>3.0.CO;2-E}{
\doi{10.1002/(SICI)1097-0258(19980430)17:8<857::AID-SIM777>3.0.CO;2-E}}.

Venables, W.N. and Ripley, B.D. (2002) \emph{Modern Applied Statistics with S.} New York: Springer.
\href{https://doi.org/10.1007/978-0-387-21706-2}{\doi{10.1007/978-0-387-21706-2}}.

Yates' continuity correction in confidence interval returned by \code{prop.test}.
\href{https://stats.stackexchange.com/questions/5206/yates-continuity-correction-in-confidence-interval-returned-by-prop-test}{
Stack Exchange}.
}
\seealso{
\code{\link[=drop_zero]{drop_zero()}}, \code{\link[stats:glm]{glm()}}, \code{\link[=odds_ratio]{odds_ratio()}},
\code{\link[stats:prop.test]{prop.test()}} and \code{\link[tibble:tibble-package]{tibble}};
\code{\link{Print_Methods}} for S3 method for printing objects of class \code{"binom_contingency"}.

Other contingency_table: 
\code{\link{contingency_table}()},
\code{\link{expl_fcts}()}
}
\concept{contingency_table}
\keyword{manip}
