% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParaAnita.R
\name{chsqfish}
\alias{chsqfish}
\title{Chi-Squared or Fisher's Exact Test}
\usage{
chsqfish(...)
}
\arguments{
\item{...}{A vector, matrix, data frame or any valid input to \code{chisq.test}}
}
\value{
A list containing the observed and expected values and the result of either \code{chisq.test()} or
\code{fisher.test()}, as appropriate.
}
\description{
Test input data using Chi-squared or Fisher's exact test as appropriate.
}
\details{
Uses \code{\link[stats]{chisq.test}()} to calculate expected values and then applies Chi-squared test if all
expected values are 5 or greater, otherwise applies \code{\link[stats]{fisher.test}()}, and
\code{\link[stats]{fisher.test}}
}
\examples{
(t <- bernoulli_data(2, 50, c(0.6, 0.4)) |>
  contingency_table(dv, iv, .rownames = TRUE))

t |> chsqfish()

(t <- bernoulli_data(3, 10, c(0.8, 0.5, 0.2)) |>
  contingency_table(dv, iv, .rownames = TRUE))

t |> chsqfish()

rm(t)
}
\seealso{
\code{\link[stats]{chisq.test}}, \code{\link{contingency_table}}.
}
