% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-glm.R
\name{comp_glm}
\alias{comp_glm}
\title{Compare Series of Nested GLMs}
\usage{
comp_glm(.data, .dep_var, ..., .family = binomial, .arrange_by = desc(AIC))
}
\arguments{
\item{.data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the binary dependent variable to be used
as the \acronym{LHS} of the model formula; should be numeric with values of \var{0} and \var{1}, or a two-column
matrix with the columns giving the numbers of successes and failures e.g., \code{cbind(pn, qn)}.}

\item{\dots}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> the \acronym{RHS} of any number of model formulae to be compared,
based on independent variables in \code{.data}.}

\item{.family}{a description of the error distribution and link function to be used in the model. Can be a character
string naming a family function, a family function or the result of a call to a family function; default
\code{binomial}.}

\item{.arrange_by}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of a column for ordering results. Use
\code{\link[dplyr:desc]{desc}} to sort by a variable in descending order; default \code{desc(AIC)}.}
}
\value{
An object of classes \code{"comp_glm"}, \code{"announce"}, inheriting from \code{\link[tibble:tibble-package]{tibble}} and hence
\code{"data.frame"}, with columns including the \acronym{RHS} of the model \code{formula}, the \code{glm} object and
eight further goodness-of-fit measures output by \code{\link[broom]{glance.glm}}, as follows: -

\item{f_rhs}{The right-hand side of the formula as supplied in the \dots argument.}

\item{.glm}{A list-column containing the glm model objects.}

\item{AIC}{Akaike's Information Criterion.}

\item{BIC}{Bayesian Information Criterion.}

\item{deviance}{Deviance of the model.}

\item{df.null}{Degrees of freedom for the null model.}

\item{df.residual}{Residual degrees of freedom.}

\item{logLik}{The log-likelihood of the model.}

\item{nobs}{Number of observations.}

\item{null.deviance}{Deviance of the null model.}
}
\description{
Compare a series of nested Bernoulli or binomial \acronym{GLMs} by supplying data, a dependent variable and
a list of terms representing the right-hand side of each of a series of model formulae.
}
\details{
\code{comp_glm()} builds the formulas from the dependent variable and the formula right-hand side, calls
\code{\link[stats:glm]{glm}}, and saves the resulting objects of class \code{"glm"} as a list column in a
\code{\link[tibble:tibble-package]{tibble}}, together with model fit information obtained using
\code{\link[broom:glance.glm]{glance.glm}} from the \pkg{\link[broom]{broom}} package. The output may be ordered by a
selected column, otherwise by default in descending order of Akaike's Information Criterion (\acronym{AIC}).

\code{comp_glm()} may be used conveniently to compare a series of related binomial \acronym{GLMs} based on different
groupings of factors added to \code{.data} using \code{\link{add_grps}}.
}
\note{
It is the user's responsibility to check models are suitably nested to ensure meaningful comparisons.
}
\examples{
## Simulate Bernoulli data

(d <- list(
        iv2 = list(g = c("a", "c", "e"), h = c("b", "d", "f")),
        iv3 = list(i = c("a", "b", "c"), j = c("d", "e", "f")),
        iv4 = list(k = c("a", "b"), l = c("c", "d"), m = c("e", "f"))
    ) |> add_grps(bernoulli_data(levels = 6), iv, .key = _))

## Models arranged in descending order of AIC by default.
d |> comp_glm(dv, iv, iv2, iv3, iv4)

## Arrange models by formula right-hand side 
(comps <- d |> comp_glm(dv, iv, iv2, iv3, iv4, .arrange_by = f_rhs))

## Inspect components of .glm list-column
lapply(comps$.glm, formula)

lapply(comps$.glm, summary)

## Convert to binomial data using binom_contingency()
(d <- d |> binom_contingency(dv))

(comps <- d |> comp_glm(cbind(pn, qn), iv, iv2, iv3, iv4))

## Inspect components of .glm list-column
lapply(comps$.glm, formula)

lapply(comps$.glm, summary)

rm(comps, d)
}
\seealso{
\code{\link{add_grps}}, \code{\link[stats:formula]{formula}}, \code{\link[broom:glance.glm]{glance.glm}} and
\code{\link[stats:glm]{glm}}.

Other comp_glm: 
\code{\link{anova_tbl}()},
\code{\link{summanov}()},
\code{\link{univ_anova}()}
}
\concept{comp_glm}
\keyword{models}
\keyword{regression}
