% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency-table.R
\name{contingency_table}
\alias{contingency_table}
\alias{xcontingency_table}
\title{Contingency Tables for Two or More Categorical Variables}
\usage{
contingency_table(.data, .dep_var, ..., .wt = NULL, .rownames = FALSE)

xcontingency_table(
  .data,
  .dep_var,
  ...,
  .crossname = NULL,
  .wt = NULL,
  .rownames = FALSE
)
}
\arguments{
\item{.data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the dependent variable, which may be a
\verb{character vector}, \code{factor}, or \code{numeric}.}

\item{\dots}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> quoted name(s) of one or more \code{factors} or
\verb{character vectors} in \code{.data}, to be included in (or excluded from) the output.}

\item{.wt}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of a numeric column in \code{.data} containing
frequency weights; default \code{NULL}.}

\item{.rownames}{\code{logical}. If \code{TRUE}, value is a data frame with the levels of the first
(or crossed) independent variable as row names, rather than a tibble; default \code{FALSE}.}

\item{.crossname}{a character string to be used as the name of the column of for the crossed variables. If
omitted, the names of the crossed variables are used combined in \dQuote{snake case}.}
}
\value{
For \code{contingency_table()}, an object of class \code{"contingency_table"}, \code{"announce"}, inheriting from
\code{\link[tibble:tibble-package]{tibble}}, or a \code{\link[base:data.frame]{data.frame}}, depending on whether
\code{.rownames = FALSE} (default) or \code{TRUE}.

Similarly for \code{xcontingency_table()}, an object of class \code{"xcontingency_table"}, \code{"announce"} inheriting from
\code{\link{tibble}} or a \code{data.frame}, again depending on the value of \code{rownames}.
}
\description{
\code{contingency_table()} compiles a contingency table for two or more categorical variables, the first of which is
typically an outcome (dependent) varable to be used for the column headings, while the remainder are typically
explanatory (independent) variables that will appear in the contingency table either as factors or optionally as row
headings.

\code{xcontingency_table()} compiles a contingency table for a categorical outcome varable and multiple categorical
explanatory variables that are \dQuote{crossed} to obtain a single explanatory factor.
}
\details{
Categorical variables (i.e. factors or character vectors) in \code{.data} required as factors in the resulting
contingency table may be selected for inclusion or exclusion using the \code{\dots} argument and the
<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax from package \pkg{\link[dplyr]{dplyr}}, including use of
\dQuote{selection helpers}. If no \code{\dots} arguments are supplied, all categorical variables in \code{.data}
(other than \code{.dep_var}) will be used.

A \link[base:list]{list} of \link[rlang:topic-defuse]{defused R expressions}, as for instance created by
\code{\link[=expl_fcts]{expl_fcts()}}, may be used as the \code{\dots} arguments and should be injected using the
\link[rlang:splice-operator]{splice-operator}, \verb{!!!}, see examples.

If \code{.wt = NULL}, the number of rows for each unique combination of the dependent and independent variables are
counted. If \code{.wt} is the quoted name of a numeric variable representing frequency weights, these are summated for
each unique combination of the dependent and independent variables.

If \code{.rownames = TRUE}, the resulting contingency table will be a conventional \code{data.frame} rather than a
\code{tibble} and the first categorical variable (other than \code{.dep_var}) will be used for row headings rather than
as a factor. Having row headings allows the result to be passed as an argument to \code{\link[stats:chisq.test]{chisq.test()}},
\code{\link[stats:fisher.test]{fisher.test()}} or \code{\link[=chsqfish]{chsqfish()}}, e.g., conveniently using \verb{|>} in a
\link[=pipeOp]{piped sequence} (see examples). However, using \code{.rownames = TRUE} for a contingency table with more than
one explanatory (independent) variable will most likely result in the error message
\dQuote{duplicate 'row.names' are not allowed}, in which case \code{xcontingency_table()} should be used instead.

Multiple categorical explanatory variables in a contingency table compiled by \code{xcontingency_table()} are
\dQuote{crossed} using \code{\link[forcats:fct_cross]{fct_cross()}}.
}
\examples{
(d <- tibble(
    iv = letters[1:4] |> sample(10, replace = TRUE),
    dv = c(0L:3L) |> sample(10, replace = TRUE)
))

d |> contingency_table(dv)

d |> contingency_table(dv, .rownames = TRUE)

## Use .data pronoun for more informative error messages
d |> contingency_table(.data$dv)

try(d |> contingency_table(dx))

try(d |> contingency_table(.data$dx))

(d <- tibble(
    iv = letters[1:4] |> sample(10, replace = TRUE) |> as.factor(),
    dv = c("Success", "Fail", "Borderline")  |> sample(10, replace = TRUE)
  ))

d |> contingency_table(dv)

d |> contingency_table(dv, .rownames = TRUE)

(d <- tibble(
    iv = letters[1:4] |> sample(100, replace = TRUE),
    dv = c("Success", "Fail", "Borderline")  |> sample(100, replace = TRUE)
  ) |> count(iv, dv))

d |> contingency_table(dv, .wt = n)

d |> contingency_table(dv, .wt = n, .rownames = TRUE) |>
    print_lf() |>
    chisq.test()

## Use .data pronoun for more informative error messages
d |> contingency_table(dv, .wt = .data$n)

try(d |> contingency_table(dv, .wt = .data$x))

rm(d)

## Using gss_cat dataset from {forcats} package
\dontshow{
  if (!requireNamespace("forcats", quietly = TRUE)) 
    warning("package 'forcats' must be installed")
  try(gss_cat <- forcats::gss_cat)
}

gss_cat |> contingency_table(race, relig, denom)

gss_cat |> contingency_table(race, !c(marital, rincome:partyid))

## Invokes warning and error message about duplicate 'row.names'
try(gss_cat |> contingency_table(race, relig, denom, .rownames = TRUE)) 

## Using xcontingency_table() avoids warning and error
gss_cat |> xcontingency_table(race, relig, denom)

gss_cat |> xcontingency_table(race, !c(marital, rincome:partyid))

gss_cat |> xcontingency_table(race, relig, denom, .crossname = "Denomination")

gss_cat |>
    xcontingency_table(race, relig, denom, .rownames = TRUE) |>
    head(10)

## Two more esoteric examples
ivars <- exprs(relig, denom)
gss_cat |> contingency_table(race, !!!ivars)

ivars <- c("relig", "denom")
gss_cat |> contingency_table(race, any_of(ivars))

rm(ivars)

\dontshow{
  rm(gss_cat)
}

}
\seealso{
\code{\link[rlang:topic-defuse]{defused R expressions}}, \code{\link[forcats:fct_cross]{fct_cross()}},
\code{\link[rlang:splice-operator]{splice-operator}} and \code{\link[tibble:tibble-package]{tibble}};
\code{\link{Print_Methods}} for S3 method for printing objects of class \code{"contingency_table"}.

Other contingency_table: 
\code{\link{binom_contingency}()},
\code{\link{expl_fcts}()}
}
\concept{contingency_table}
\keyword{manip}
