% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast-names.R
\name{contr_colnames}
\alias{contr_colnames}
\alias{contr_colnames<-}
\alias{contr_colpfx<-}
\title{Get and Set Contrast Matrix Column Names}
\usage{
contr_colnames(x)

contr_colnames(x) <- value

contr_colpfx(x) <- value
}
\arguments{
\item{x}{a \code{factor} for which the contrast column headings are to be set or viewed.}

\item{value}{a character vector containing the contrast column names, of length equal to the number of contrasts;
or in the case of \code{contr_colpfx()<-} a character or string used to prefix the existing contrast column names.}
}
\description{
Set and view the column names of contrasts associated with a factor.
}
\details{
\code{contr_colnames()} returns the current column names of the contrasts for a factor.

\code{contr_colnames()<-} sets the column names of the contrasts for a factor.

\code{cntr_pfx()<-} prefixes the current column names of the contrasts for a factor with the character or string
provided. This can be useful when factor levels are elided with the factor name as, for instance, in the printed output of
\code{\link[stats]{summary.glm}}.

If contrasts are not set for \var{x}, both \code{contr_colnames()<-} and \code{cntr_pfx()<-} set the contrast attribute
using the default function from \code{\link[base]{options}("contrasts")} before modifying the column names.
}
\examples{
d <- data.frame(
  f = gl(5, 5, labels = LETTERS[1:5]),
  dv = sample(c(0,1), 25, replace = TRUE)
) |> print()

contrasts(d$f) <- contr.helmert
contrasts(d$f)
contr_colnames(d$f)
glm(dv ~ f, family = binomial, data = d) |> summary()

contr_colnames(d$f) <- c("A v. B", "AB v. C", "ABC v. D", "ABCD v. E")
contr_colnames(d$f)
contrasts(d$f)
glm(dv ~ f, family = binomial, data = d) |> summary()

contr_colpfx(d$f) <- ": "
contr_colnames(d$f)
glm(dv ~ f, family = binomial, data = d) |> summary()

rm(d)
}
\seealso{
\code{\link[stats]{contrast}}, \code{\link[stats]{contrasts}} and \code{\link[base]{factor}}.

Other contrast-names: 
\code{\link{helm_names}()}
}
\concept{contrast-names}
