% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-manip.R
\name{fct_to_num}
\alias{fct_to_num}
\title{Factor as Numeric}
\usage{
fct_to_num(f)
}
\arguments{
\item{f}{factor to be converted to numeric values}
}
\value{
Numeric
}
\description{
Transform a factor to approximately its original numeric values.
}
\details{
See \sQuote{Warning} section of \code{\link[base]{factor}}: \dQuote{In particular, \code{as.numeric}
applied to a factor is meaningless, and may happen by implicit coercion. To transform a factor
\code{f} to approximately its original numeric values, \code{as.numeric(levels(f))[f]} is
recommended and slightly more efficient than \code{as.numeric(as.character(f))}.}
}
\examples{
f <- factor(2001:2020)

f

f |> as.numeric()	# Returns codes for factor levels, not what is required

f |> fct_to_num()	# Returns approximate numeric values, as required

rm(f)

}
\seealso{
\code{\link[base]{factor}}

Other factor-manip: 
\code{\link{add_grps}()}
}
\concept{factor-manip}
\keyword{category}
\keyword{math}
