% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-contrasts.R
\name{get_contr_data}
\alias{get_contr_data}
\alias{set_contr_treat}
\alias{set_contr_treat<-}
\title{Get and Set Treatment Contrasts for Independent Variables in Data}
\usage{
get_contr_data(data, ...)

set_contr_treat(data, ..., .baseline = NULL, .verbose = TRUE)

set_contr_treat(data, ..., .verbose = FALSE) <- value
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{\dots}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> The selection of one or more factors in \code{data} for
getting or setting contrasts.}

\item{.baseline}{a \code{numeric} vector of \code{length} equal to the number of contrasts to be set specifying which
level is considered the baseline.}

\item{.verbose}{\code{logical}, whether to print "before and after" contrast matrices for factors in \code{data}.}

\item{value}{\code{numeric}, see \code{.baseline} argument.}
}
\value{
The original dataframe or \code{\link[tibble:tibble-package]{tibble}} with the \code{"contrasts"} attributes set for
selected \code{factors}.
}
\description{
\code{get_contr_data()} shows the \code{"contrasts"} attributes of all or selected factors within a data frame.

\code{set_contr_treat()} sets the \code{"contrasts"} attribute for selected factors within a data frame to a
treatment contrast matrix with individually specified baseline levels.

\code{set_contr_treat()<-} is the replacement function form.
}
\details{
\code{get_contr_data()} prints the \code{"contrasts"} attributes of all or selected factors and returns \code{data}
invisibly.

Factors in \code{.data} may be selected for getting and setting contrasts using the \code{\dots} argument with
the <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax of package \pkg{dplyr}, including use of
\strong{selection helpers}.  If no \code{\dots} arguments are supplied, all categorical variables in \code{data} (i.e.,
\code{character} or \code{factor} columns) will be selected.

The \code{"contrasts"} attribute of \code{factors} selected with \code{\dots} are set using the contrast function
\code{\link[stats:contrast]{contr.treatment}} with baseline factor levels as specified numerically in the argument
\code{.baseline} or the \code{value} argument in the case of the replacement function form \code{set_contrasts()<-}.
If no \code{.baseline} argument is supplied, by default the first factor level is used as baseline.

The individual \code{.baseline} (or \code{.value}) argument values are capped to be no greater than
\code{\link[base:nlevels]{nlevels}} for each of the corresponding factors selected in \code{...}. Hence, to ensure the last
level is the reference level, a baseline value can be specified as a large integer (e.g., \var{99}L), which may be
convenient when using \code{set_contr_treat()} programmatically.
}
\examples{
# Coming soon!

}
\seealso{
\code{\link[stats:contrasts]{contrasts}}, \code{\link[stats:zC]{C}}, \code{\link[stats:contrast]{contr.treatment}}.

Other get-contrasts: 
\code{\link{get_contrasts}()}
}
\concept{get-contrasts}
\keyword{array}
\keyword{design}
\keyword{regression}
