% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-model.R
\name{glm_plotdata}
\alias{glm_plotdata}
\alias{glm_plotdata.binom_contingency}
\alias{glm_plotdata.data.frame}
\alias{glm_plotdata.default}
\title{Collate Data for Plotting Univariable GLM Predictions with Error Bars}
\usage{
glm_plotdata(object, ...)

\method{glm_plotdata}{binom_contingency}(
  object,
  ...,
  .ind_var,
  .ungroup = NULL,
  conf_level = 0.95,
  type = c("link", "response")
)

\method{glm_plotdata}{data.frame}(
  object,
  ...,
  .dep_var,
  .ind_var,
  .ungroup = NULL,
  conf_level = 0.95,
  type = c("link", "response")
)

\method{glm_plotdata}{default}(object, ..., conf_level = 0.95, type = c("link", "response"))
}
\arguments{
\item{object}{an object from which the data for plotting univariable GLM predictions are to be collated; may be a
\link[=binom_contingency]{binomial contingency table}, a \link[base:data.frame]{data frame} (or a data frame extension e.g.,
a \code{\link[tibble:tibble-package]{tibble}}), or a \code{\link[stats:glm]{glm}} as used by the default S3 method.}

\item{\dots}{further arguments passed to or from other methods.}

\item{.ind_var}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the independent variable.}

\item{.ungroup}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the column containing the ungrouped levels
of \code{.ind_var}, see details; default \code{NULL}.}

\item{conf_level}{the confidence level required for the error bars; default \var{0.95}. If \code{NA}, error bars are
standard error.}

\item{type}{the type of prediction required. The default is on the scale of the linear predictors;
the alternative \code{"response"} is on the scale of the response variable; default \code{"link"}.}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name(s) of the response variable(s) in the data
representing the number of successes and failures respectively, see \code{\link[stats:glm]{glm()}}; default
\code{cbind(pn, qn)}.}
}
\value{
An object of class \code{"glm_plotdata"}, \code{"announce"})\verb{, inheriting from [}tibble\verb{][tibble::tibble-package], with values on the linear predictor or response scale (depending on }type`) in columns as follows: -

\item{level}{Level of the independent variable.}

\item{grouped}{Grouped levels of the independent variable.}

\item{n}{Number of observations.}

\item{obs}{Observed values.}

\item{pred}{Values predicted by the model.}

\item{lower}{Lower extent of error bar.}

\item{upper}{Upper extent of error bar.}

It also has attributes \code{"conf_level"}, signifying the confidence level, \code{"subtitle"}, by default the name of the
independant variable, and \code{"type"} (see argument \code{type}).
}
\description{
\code{glm_plotdata()} outputs data collated suitably for easy creation of standardised plots with error bars representing
confidence intervals or standard errors, based on predictions from univariable general linear models (\acronym{GLM}s).
}
\details{
This function works with univariable \acronym{GLM}s of Bernoulli or binomial data having a dependent variable
expressed as numbers of successes and failures and a single independent variable with multiple levels and its output
may be plotted conveniently using the \code{\link[ggplot2:ggplot]{ggplot()}} S3 method
\code{\link[=ggplot.glm_plotdata]{ggplot.glm_plotdata()}}.

\code{glm_plotdata()} allows exploration of proposed groupings of the levels of the independent variable, such as
obtained using \code{\link[=add_grps]{add_grps()}} or \code{\link[forcats:fct_collapse]{fct_collapse()}}, and will include both the
grouped and ungrouped levels in its output. In such cases, \code{.ind_var} should contain the groupings and the \code{.ungroup}
argument should name a column in \code{object}'s data containing the ungrouped levels, see examples. The grouped levels are
used as the independent variable in the \acronym{GLM} and are shown within the output object in the column \code{grouped}
while the ungrouped levels are shown in the column \code{level}. If the \code{.ungroup} is \code{NULL} (the default), levels of
\code{.ind_var} will appear in the column \code{level} and the \code{grouped} column in the output will contain \code{\link[base:NA]{NA}}.

If \code{conf_level} is a value such as \var{0.95} (the default) or a similar value, \code{lower} and \code{upper}
values in the output delimit the prediction confidence intervals at that confidence level. If \code{conf_level} is
\code{\link[base:NA]{NA}}, then the \code{lower} and \code{upper} are the model predictions ±standard error.

If \code{type = "link"}, then the linear predictors and their confidence intervals or ±standard errors are obtained.
If \code{type = "response"}, then the linear predictors and their confidence intervals or ±standard errors are
transformed back to the response scale using the link inverse function.
}
\note{
Confidence intervals are calculated from the standard errors of the parameter estimates using the quantiles of
the t distribution with \emph{n - 1} degrees of freedom, at the probability given by \code{conf_level}. These
confidence intervals are generally more conservative i.e., a little wider, than those obtained by "profiling"
(e.g., using \code{\link[MASS:confint]{confint.glm}}). If the \code{conf_level} argument is \code{\link[base:NA]{NA}}, standard
error is shown rather than a confidence interval.
}
\examples{
(d <- binom_data())

## Ungrouped data
## 95\% Confidence interval by default
d |> glm_plotdata(.ind_var = iv)                     ## Linear predictor scale
d |> glm_plotdata(.ind_var = iv, type = "response")  ## Response scale

## Standard error
d |> glm_plotdata(.ind_var = iv, conf_level = NA)                     ## Linear predictor scale
d |> glm_plotdata(.ind_var = iv, conf_level = NA, type = "response")  ## Response scale

## Grouped data
(d <- list(iv2 = list(ab = c("a", "b"), cd = c("c", "d"))) |>
    add_grps(d, iv, .key = _))

## 95\% Confidence interval by default
d |> glm_plotdata(.ind_var = iv2, .ungroup = iv)                     ## Linear predictor scale
d |> glm_plotdata(.ind_var = iv2, .ungroup = iv, type = "response")  ## Response scale

## Standard error on linear predictor scale
d |> glm_plotdata(.ind_var = iv2, .ungroup = iv, conf_level = NA)

## Standard error on response scale
d |> glm_plotdata(.ind_var = iv2, .ungroup = iv, conf_level = NA, type = "response")

rm(d)
}
\seealso{
\code{\link{add_grps}}, \code{\link{binom_contingency}}, \code{\link[stats:glm]{glm}},
\code{\link[purrr:imap]{imap}} and  \code{\link[tibble:tibble-package]{tibble}}.

Other plot_model: 
\code{\link{Plot_Model}},
\code{\link{glm_plotlist}()},
\code{\link{var_labs}()}
}
\concept{plot_model}
\keyword{dplot}
