% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-model.R
\name{glm_plotlist}
\alias{glm_plotlist}
\title{Data for Plotting Univariable GLM Predictions and Error Bars for Multiple Independent Variables}
\usage{
glm_plotlist(
  data,
  .dep_var,
  ...,
  .ungroups = NULL,
  .conf_level = 0.95,
  .type = c("link", "response"),
  .facet_by = NULL
)
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the \code{numeric} response variable in
the data, either a vector with values of \code{1} and \code{0}, representing the number of successes and failures
respectively, or as a two-column matrix with the columns giving the numbers of successes and failures see
\code{\link[stats:glm]{glm()}}.}

\item{\dots}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> independent variables to be included in the plot data.}

\item{.ungroups}{a named character vector of ungrouped levels of independent variables specified in \code{.ind_var},
see details; default \code{NULL}.}

\item{.conf_level}{the confidence level required for the error bars; default \var{0.95}. If \code{NA}, error bars are
standard error.}

\item{.type}{the type of prediction required. The default is on the scale of the linear predictors;
the alternative \code{"response"} is on the scale of the response variable; default \code{"link"}.}

\item{.facet_by}{\code{NULL}, the default; or, if the output is to be combined into a single object to be used for a
faceted plot, a \verb{character vector} of length one used to name an additional column containing the names of the
independent variables.}
}
\value{
If the argument \code{.facet_by} is \code{NULL}, a \code{list} of \code{"glm_plotdata"} objects suitable for producing multiple plots
using \code{\link[ggplot2:ggplot]{ggplot()}}. Otherwise, a single \code{"glm_plotdata"} object with an additional column taking
its name from \code{.facet_by} and containing the names of the independent variables.
}
\description{
\code{glm_plotlist()} formats data for plotting univariable \acronym{GLM} predictions with error bars
for each of a number of independent variables.
}
\details{
\code{glm_plotlist()} invokes \code{\link[=glm_plotdata]{glm_plotdata()}} to create a \code{list} of \code{"glm_plotdata"} objects for
plotting univariable \acronym{GLM} predictions with error bars for each of a number of independent variables
in \code{data}. Independent variables to be included are selected using the \code{\dots} argument with the
<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax of package \pkg{\link[dplyr]{dplyr}}, including use of
\dQuote{selection helpers}.

Like \code{\link[=glm_plotdata]{glm_plotdata()}}, \code{glm_plotlist()} allows exploration of proposed groupings of levels of
independent variables (e.g. as obtained using \code{\link[=add_grps]{add_grps()}} or \code{\link[forcats:fct_collapse]{fct_collapse()}})
and inclusion of both grouped and ungrouped levels in the \code{"glm_plotdata"} objects comprising its output list. In
such cases, the \code{.ungroups} argument is used to provide a named \verb{character vector} of the names of the corresponding
factors in \code{data} giving the grouped and ungrouped levels of the form \code{ungrouped_name = "grouped_name"}; levels not
otherwise mentioned will be left as is.

The grouped levels are used as the independent variable in the \acronym{GLM} invoked by \code{glm_plotdata()} and are
output in the column \code{grouped} within the corresponding \code{"glm_plotdata"} object, while the ungrouped levels are shown
in the column \code{level}, see \code{\link[=glm_plotdata]{glm_plotdata()}}.

The \code{.conf_level} and \code{.type} arguments are handled as for \code{\link[=glm_plotdata]{glm_plotdata()}}.

\code{\link[=glm_plotlist]{glm_plotlist()}} may be used in conjunction with \code{\link[base:lapply]{lapply()}} (or
\pkg{\link[purrr]{purrr}} package \link[purrr:map]{map()}) to rapidly obtain multiple plots of univariable
\acronym{GLM}s for a number of independent variables.

Levels of independent variables for which the observed values are all zero or all one are not included in the output,
although they are taken into consideration in calculating denominators in the case of grouped levels.
}
\examples{
# Coming soon!

}
\seealso{
\code{\link[=add_grps]{add_grps()}}, \code{\link[dplyr:bind_rows]{bind_rows()}} and \code{\link[forcats:fct_collapse]{fct_collapse()}}.

Other plot_model: 
\code{\link{Plot_Model}},
\code{\link{glm_plotdata}()},
\code{\link{var_labs}()}
}
\concept{plot_model}
\keyword{dplot}
