% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast-names.R
\name{helm_names}
\alias{helm_names}
\alias{helm_names<-}
\title{Create and Set Names for Helmert Contrasts}
\usage{
helm_names(x)

helm_names(x) <- value
}
\arguments{
\item{x}{a \code{factor} for which the contrast column headings are to be named.}

\item{value}{a character vector or \code{\link{list}} of length \var{two} (any subsequent elements will be ignored),
which are separators used in creating the headings; the first is a "within" separator and the second a "between"
separator, see examples.}
}
\description{
Create and set column names for Helmert contrasts associated with a factor.
}
\details{
Helmert contrasts for a factor contrast the second level with the first, the third with the average of the first two,
and so on. A contrasts matrix can be set as an attribute of a factor using the function
\code{\link[stats:contrasts]{contrasts<-}}, and using the contrasts function \code{\link[stats:contrast]{contr.helmert}} as an
argument i.e., \code{contrasts(f)<- contr.helmert}, will create and set an appropriate Helmert contrasts matrix
attribute. However the columns of the resulting Helmert contrasts matrix are unnamed. \code{helm_names()<-} creates
these column names based on the names the levels of the factor
\var{x}.

If factor \var{x} does not have an associated Helmert contrast matrix, \code{helm_names()<-} will set one, giving a
warning. The names of the levels of factor \var{x} need to be fairly short for function this to be helpful.

The function \code{helm_names()} is an alias for \code{\link{contr_colnames}} and included simply for completeness
and to optimise help searches.
}
\examples{
f <- gl(5, 5, labels = LETTERS[1:5]) |> print()

contrasts(f)<- contr.helmert
f
helm_names(f)

helm_names(f)<- c(":", "v")
f
helm_names(f)

contrasts(f)<- NULL
f

helm_names(f)<- c("-", "vs.")
f
helm_names(f)

rm(f)
}
\seealso{
\code{\link[stats:contrast]{contrast}}, \code{\link[stats:contrasts]{contrasts}} and \code{\link[stats:contrast]{contr.helmert}}.

Other contrast-names: 
\code{\link{contr_colnames}()}
}
\concept{contrast-names}
