% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_data.R
\name{levels_data}
\alias{levels_data}
\alias{nlevels_data}
\title{Levels of all Factors in Data}
\usage{
levels_data(data)

nlevels_data(data)
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}
}
\value{
\code{levels_data()} returns a named \code{list} comprising the \code{levels} of each \code{factor} in \code{data}. \code{nlevels_data()}
returns a named \verb{integer vector} comprising the number of \code{levels} of each \code{factor} in \code{data}.
}
\description{
\code{levels_data()} returns the levels for all factors in data.

\code{nlevels_data()} returns the number of levels for all factors in data.
}
\details{
Does what it says on the tin.
}
\examples{
## Using gss_cat dataset from {forcats} package
\dontshow{
   if (!requireNamespace("forcats", quietly = TRUE)) 
       warning("package 'forcats' must be installed")
   try(gss_cat <- forcats::gss_cat)
}

gss_cat |> names()
gss_cat |> levels_data()
gss_cat |> nlevels_data()

\dontshow{
    rm(gss_cat)
}
}
\seealso{
\code{\link[base:levels]{levels}}.

Other levels_data: 
\code{\link{good_levels}()}
}
\concept{levels_data}
