% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{lf}
\alias{lf}
\alias{print_lf}
\title{Pipe-Friendly Line Feeds and Printing}
\usage{
lf(x, n = 1)

print_lf(x, n = 1)
}
\arguments{
\item{x}{Object to be piped.}

\item{n}{Number of line feeds; default \code{1}.}
}
\value{
Invisibly returns its first argument.
}
\description{
\code{lf()} outputs one or more line feeds during a piped sequence.
}
\details{
\code{print_lf()} prints an object in a piped sequence then outputs one or more line feeds.

An object passed as argument in a piped sequence is printed and/or one or more line feeds are
output during a piped sequence using \code{\link[base]{cat}()}. This can be useful to separate
lines of printed output, see examples.
}
\examples{
obj <- "Lorum ipsum dolor sit amet"
obj |> lf()               # line feed, object returned invisibly
obj |> lf(3)              # three line feeds, object returned invisibly
(obj |> lf(3))            # three line feeds, returned object rendered visible
obj |> lf(3) |> paste("consectetur adipiscing elit", sep = ", ")

obj |> print() |> lf(3)   # line feeds are unexpectedly before printed output.

## Use print_lf() instead
obj |> print_lf()         # object printed with line feed and returned invisibly
obj |> print_lf(3)        # object printed with three line feeds and returned invisibly
(obj |> print_lf(3))      # Ditto, then rendered visible
obj |> print_lf(3) |> paste("consectetur adipiscing elit", sep = ", ")

rm(obj)
}
\seealso{
\code{\link[base:cat]{cat}}.

Other print: 
\code{\link{announce}()},
\code{\link{print_all}()}
}
\concept{print}
\keyword{print}
