% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds-ratio.R
\name{odds_ratio}
\alias{odds_ratio}
\alias{odds_ratio.binom_contingency}
\alias{odds_ratio.data.frame}
\alias{odds_ratio.default}
\alias{print.odds_ratio}
\alias{print_all.odds_ratio}
\title{Odds Ratios, Standard Errors, Confidence Intervals and P-Values for Binomial GLMs}
\usage{
odds_ratio(object, ...)

\method{odds_ratio}{binom_contingency}(
  object,
  ...,
  .ind_var,
  .level = 0.95,
  .printcall = FALSE,
  .stat = FALSE,
  .print_contr = FALSE
)

\method{odds_ratio}{data.frame}(
  object,
  ...,
  .dep_var = cbind(pn, qn),
  .ind_var,
  .level = 0.95,
  .printcall = FALSE,
  .stat = FALSE,
  .print_contr = FALSE
)

\method{odds_ratio}{default}(
  object,
  ...,
  .level = 0.95,
  .printcall = TRUE,
  .stat = FALSE,
  .print_contr = FALSE
)

\method{print}{odds_ratio}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{print_all}{odds_ratio}(x, linefeeds = NULL)
}
\arguments{
\item{object}{an object from which the odds ratios are to be calculated, which may be a
\link[=binom_contingency]{binomial contingency table}, a \link[base:data.frame]{data frame} (or a data frame extension e.g.,
a \code{\link[tibble:tibble-package]{tibble}}), or a \code{\link[stats:glm]{glm}}.}

\item{\dots}{further arguments passed to or from other methods.}

\item{.ind_var}{<\code{\link[rlang:args_data_masking]{data-masked}}> quoted name of an independent variable, which may be
either a character vector or factor.}

\item{.level}{the confidence level required; default \var{0.95}.}

\item{.printcall}{\code{logical}, whether or not to print the call for the \acronym{GLM}.}

\item{.stat}{\code{logical}, whether or not to print \code{z} or \code{t} statistic for the \acronym{GLM};
default \code{FALSE}.}

\item{.print_contr}{\code{logical}. If \code{TRUE}, and \code{.ind_var} has a contrast attribute set, the contrast
matrix will be printed; default \code{FALSE}.}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masked}}> quoted name(s) of the response variable(s) in the data
representing the number of successes and failures respectively, see \code{\link[stats:glm]{glm()}}; default
\code{cbind(pn, qn)}.}
}
\value{
An object of classes \code{"odds_ratio"}, \code{"announce"}, inheriting from\code{"data.frame"} possibly also from \code{"tbl"}
and \code{"tbl_df"}, and containing the following columns: -

\item{parameter}{The names of the model parameters.}

\item{estimate}{The estimate of the regression coefficient.}

\item{se}{The standard error of the estimate.}

\item{z (or t) value}{Optionally, the value of the z (or t) statistic for the estimate.}

\item{p_val}{The p-value for the estimate.}

\item{odds_ratio}{The odds ratio.}

\item{ci}{The lower and upper confidence intervals for the odds ratio, by default at the 2.5\% and 97.5\% levels.}

\item{sig}{Stars for statistical significance.}
}
\description{
\code{odds_ratio()} calculates odds ratios and their profiled confidence intervals for \acronym{GLMs} and outputs these
together with the estimates of the regression coefficients, their standard errors and probabilities.
}
\details{
\code{odds_ratio()} is a generic function used to calculate odds ratios and their profiled confidence intervals for
univariable \acronym{GLMs} with a single categorical independent variable, or for multivariable \acronym{GLMs}, and
output these together with the estimates of the regression coefficients, their standard errors and probabilities.
The function invokes particular \code{\link[utils:methods]{methods}} which depend on the \code{\link[base:class]{class}} of the first
argument.

The default S3 method is for objects of class \code{"glm"} returned by \code{\link[stats:glm]{glm()}} and can be used with
unvariable \acronym{GLMs}, or with multivariable \acronym{GLMs} to calculate "adjusted" odds ratios. Optionally,
if \code{printcall = TRUE} the original call to \code{\link[stats:glm]{glm()}} may be retrieved from the \code{"glm"} object supplied
as an argument and printed.

Currently, the S3 methods for classes \code{"data.frame"} and \code{"binom_contingency"} can only be used with univariable
\acronym{GLMs}.

If \code{.print_contr = TRUE} and any \code{factor} independent variables have a contrast \code{\link[base:attributes]{attribute}} set,
the \code{\link[stats:contrast]{contrast matrix}} will be printed. Contrasts may be set conveniently for \code{factors} in
\code{data} using \code{\link[=set_contrasts]{set_contrasts()}}, see examples.

Confidence intervals for odds ratios are based on profile likelihood and calculated using
\code{\link[MASS:confint]{confint.glm()}}. The confidence level may be adjusted using \code{.level}, otherwise the default
value of \var{0.95} is used.
}
\examples{
## Create simulated Bernoulli data
(d <- bernoulli_data())

## Invoking the S3 method for class "data.frame" and using the default
## contrasts from options("contrasts")
## — contrasts not printed 
d |> odds_ratio(.dep_var = dv, .ind_var = iv)

## Using the default contrasts from options("contrasts")
## — adjust confidence level, contrasts are printed 
d |> odds_ratio(.dep_var = dv, .ind_var = iv, .level = 0.99, .print_contr = TRUE)

## Specifying treatment contrasts, with last level as base
## — contrasts not printed
d |> set_contrasts(iv, base =  99L, contr = contr.treatment) |>
    odds_ratio(.dep_var = dv, .ind_var = iv)

## Specifying treatment contrasts, with last level as base
## — contrasts printed
d |> set_contrasts(iv, base =  99L, contr = contr.treatment) |>
    odds_ratio(.dep_var = dv, .ind_var = iv, .print_contr = TRUE)

## Helmert contrasts specified
## — contrasts printed
d |> set_contrasts(iv, contr = contr.helmert) |>
    odds_ratio(.dep_var = dv, .ind_var = iv, .print_contr = TRUE)

# Set default unordered contrasts in options("contrasts") to Helmert
options("contrasts" =  c(unordered = "contr.helmert", ordered = "contr.poly"))
getOption("contrasts")

## Using the default, unordered Helmert contrasts
## — contrasts printed
d |> odds_ratio(.dep_var = dv, .ind_var = iv, .print_contr = TRUE)

## Specify treatment contrasts
## — contrasts printed
d |> set_contrasts(iv, contr = contr.treatment) |>
    odds_ratio(.dep_var = dv, .ind_var = iv, .print_contr = TRUE)

# Restore default contrasts in options("contrasts")
options("contrasts" =  c(unordered = "contr.treatment", ordered = "contr.poly"))
options("contrasts")

## Invoking the S3 method for class "binom_contingency" 
d |> binom_contingency(dv, iv) |> odds_ratio(.ind_var = iv)

## Create multivariable glm object and specify treatment contrasts
(d <- list(
    iv2 = list(g = c("a", "c", "e"), h = c("b", "d", "f")),
    iv3 = list(i = c("a", "b", "c"), j = c("d", "e", "f"))
) |> add_grps(binom_data(levels = 6), iv, .key = _))

set_contr_treat(d, num_range("iv", 2:3)) <- c(1L, 2L)
get_contr_data(d)

glm1 <- glm(cbind(pn, qn) ~ iv2 + iv3, family = binomial, data = d)

glm1 |> summary()

## Invoking the S3 method for class "glm"
glm1 |> odds_ratio()

glm1 |> odds_ratio(.printcall = F, .stat = TRUE, .print_contr = TRUE)

## Compare S3 method for class "glm" to that for "data.frame"
## — only possible for univariable analyses
(d <- binom_data())

glm(cbind(pn, qn) ~ iv, family = binomial, data = d) |>
    odds_ratio()

d |> odds_ratio(.ind_var = iv)

## Helmert contrasts given more easily readable names
d |> set_contrasts(iv) <- contr.helmert
helm_names(d$iv) <- c(":", "v")
d |> get_contrasts(iv)

## Add separator as last little tweak ;-)
contr_colpfx(d$iv) <- ": "

glm(cbind(pn, qn) ~ iv, family = binomial, data = d) |>
    odds_ratio()

d |> odds_ratio(.ind_var = iv)

## Printing lengthier output with print_all()
binom_data(26, 100) |>
    odds_ratio(.ind_var = iv, .print_contr = TRUE) |>
    print_all()

rm(d, glm1)
}
\seealso{
\code{\link[=binom_contingency]{binom_contingency()}}, \code{\link[stats:contrast]{contrast matrix}},
\code{\link[stats:contrasts]{contrasts()}}, \code{\link[stats:glm]{glm()}}, \code{\link[=set_contrasts]{set_contrasts()}},
\code{\link[stats:summary.glm]{summary.glm()}} and \code{\link[=print_all]{print_all()}} for easy printing of all the rows of
a longer \code{tibble}.
}
\concept{odds-ratio}
\keyword{models}
\keyword{regression}
