% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-dplyr.R, R/reexport-forcats.R,
%   R/reexport-ggplot2.R, R/reexport-purrr.R, R/reexport-rlang.R,
%   R/reexport-stringr.R, R/reexport-tibble.R, R/reexport-tidyselect.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{count}
\alias{mutate}
\alias{rename}
\alias{fct_count}
\alias{fct_recode}
\alias{as_labeller}
\alias{element_text}
\alias{ggplot}
\alias{labs}
\alias{theme_get}
\alias{theme_set}
\alias{theme_update}
\alias{list_rbind}
\alias{list_transpose}
\alias{map2}
\alias{map2_chr}
\alias{map_chr}
\alias{map_dbl}
\alias{map_int}
\alias{\%@\%}
\alias{\%|\%}
\alias{\%||\%}
\alias{eval_tidy}
\alias{exprs}
\alias{set_names}
\alias{str_to_title}
\alias{tibble}
\alias{all_of}
\alias{any_of}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{last_col}
\alias{matches}
\alias{num_range}
\alias{starts_with}
\alias{where}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}}

  \item{forcats}{\code{\link[forcats]{fct_count}}, \code{\link[forcats]{fct_recode}}}

  \item{ggplot2}{\code{\link[ggplot2]{as_labeller}}, \code{\link[ggplot2:element]{element_text}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2]{theme_get}}, \code{\link[ggplot2:theme_get]{theme_set}}, \code{\link[ggplot2:theme_get]{theme_update}}}

  \item{purrr}{\code{\link[purrr:list_c]{list_rbind}}, \code{\link[purrr]{list_transpose}}, \code{\link[purrr:map]{map_chr}}, \code{\link[purrr:map]{map_dbl}}, \code{\link[purrr:map]{map_int}}, \code{\link[purrr]{map2}}, \code{\link[purrr:map2]{map2_chr}}}

  \item{rlang}{\code{\link[rlang:op-get-attr]{\%@\%}}, \code{\link[rlang:op-na-default]{\%|\%}}, \code{\link[rlang:op-null-default]{\%||\%}}, \code{\link[rlang]{eval_tidy}}, \code{\link[rlang:defusing-advanced]{exprs}}, \code{\link[rlang]{set_names}}}

  \item{stringr}{\code{\link[stringr:case]{str_to_title}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect:all_of]{any_of}}, \code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{where}}}
}}

