% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-objects.R
\name{rm_objects}
\alias{rm_objects}
\title{Remove Sequentially Numbered Objects from Workspace}
\usage{
rm_objects(basename, suffixes, envir = parent.frame())
}
\arguments{
\item{basename}{Common base name (quoted) of the series of objects.}

\item{suffixes}{A numeric or character vector representing the suffixes of the series of objects.}

\item{envir}{An environment from which to remove objects. Use \code{.GlobalEnv} for the workspace; default
\code{\link[base:sys.parent]{parent.frame()}} i.e., the environment in which \code{rm_objects()} was called.}
}
\value{
A character vector of matching names remaining in the calling \code{\link[base:environment]{environment}}, usually
the workspace unless \code{rm_objects()} was called within a function, or another specified \code{environment}, returned
invisibly.
}
\description{
Remove a series of sequentially named objects from the workspace or from another specified
\code{environment}. For example, conveniently remove a series of sequentially numbered models.
}
\details{
\code{rm_objects()} lists all objects in the workspace (or another specified  \code{\link[base:environment]{environment}})
whose names start with \code{basename}, then removes any in which \code{basename} is followed by an element included in
\code{suffixes}, and finally lists all remaining objects with names matching \code{basename}.

\acronym{#GF}
}
\examples{
 ## Note: running outside example() will be more informative

 ## Create some sequentially numbered objects
 model1 <- model2 <- model3 <- model4 <- lm(1~1)
 ls(pattern = "model")

 ## Remove three of them
 rm_objects(model, 1:3)

 ## Create some sequentially named objects
 model_a <- model_b <- model_c <- model_d <- lm(1~1)
 ls(pattern = "model_")

 ## Remove three of them
 rm_objects(model_, letters[1:3])

 ## Use within a function
 (\() {                  ## Anonymous function, but doesn't have to be
   model1 <- model2 <- model3 <- model4 <- model5 <- lm(1~1)
   rm_objects(model, 1:5)
 })()

 ls(pattern = "model")

 rm_objects(model, c(4, "_d"))
}
\seealso{
\code{\link[base:environment]{environment}}, \code{\link[base:ls]{ls()}} and \code{\link[base:rm]{rm()}}.
}
\keyword{environment}
