% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{Simulate_Data}
\alias{Simulate_Data}
\alias{bernoulli_data}
\alias{binom_data}
\title{Simulated Bernoulli and  Binomial Proportion Data}
\usage{
bernoulli_data(
  levels = 5,
  length = 66,
  probs = seq(0.5, 0.1, length.out = levels)
)

binom_data(
  levels = 5,
  length = 66L,
  probs = seq(0.5, 0.1, length.out = levels)
)
}
\arguments{
\item{levels}{numeric, the desired number of levels of the independent variable \code{iv};  default \var{5}.}

\item{length}{numeric, the desired number of simulated observations per level of the independent variable \code{iv};
default \var{20}.}

\item{probs}{a numeric vector of the same length as levels, representing the probabilities of success for each
corresponding level; default \code{seq(0.5, 0.1, length.out = levels)}.}
}
\value{
An object of class \code{"announce"} inheriting from \code{\link[tibble:tibble-package]{tibble}} with column \code{iv} for the independant
variable, and for \code{bernoulli_data()}, column \code{dv} representing the dependant variable; and for \code{binom_data()}, columns
\code{pn} and \code{qn} representing the number of "successes" and "failures", as follows: -

\item{\code{iv}}{a \code{factor} representing levels of the independant variable.}

\item{\code{dv}}{an \code{integer} representing the value of the dependent variable.}

\item{\code{pn}}{an \code{integer} representing  the number of successes.}

\item{\code{qn}}{an \code{integer} representing  the number of failures.}
}
\description{
\code{bernoulli_data()} creates a simulated univariate Bernoulli data set having a dependent variable \code{dv} with values of
\var{0} and \var{1}, and an independent variable \code{iv} with levels represented by lower case letters.

\code{binom_data()} creates a simulated univariate binomial proportion data set having a variable \code{pn} representing the
number of successes, a variable \code{qn} representing the number of failures, and an independent variable \code{iv} with
levels represented by lower case letters.
}
\details{
A random sample from a Bernoulli distribution is obtained for each level of the independent variable \code{iv}, at the
corresponding probability given in \code{probs}, using \code{\link[stats:Binomial]{rbinom()}} with \code{size = 1}. The result is returned
as a \code{\link[tibble:tibble-package]{tibble}} with two columns, \code{iv} representing the level of the independent variable and
\code{dv} representing the simulated data. The result may be easily converted into (simulated) proportion data and inspected
using \code{\link[=binom_contingency]{binom_contingency()}}, see examples.

A random sample from a binomial distribution of size \code{length} is obtained for each level of the independent variable
\code{iv}, at the corresponding probability given in \code{probs}, using \code{\link[stats:Binomial]{rbinom()}} with \code{size = levels}.

\code{bernoulli_data()} and \code{binom_data()} are used for demonstrating and testing functions such as
\code{\link[=contingency_table]{contingency_table()}}, \code{\link[=binom_contingency]{binom_contingency()}} and \code{\link[=odds_ratio]{odds_ratio()}}.
}
\note{
The default length of 66 is the minimum number of trials with probability of success of 0.1 for which the
overall probability of zero failures is less than 1 in 1000 i.e., \eqn{(1 - 0.1)^{66} < 0.001}.
}
\examples{
bernoulli_data()
bernoulli_data() |> binom_contingency(dv, iv)
bernoulli_data(probs = seq(0.4, 0, length.out = 5)) |> binom_contingency(dv, iv)

binom_data()
binom_data(probs = seq(0.4, 0, length.out = 5))

}
\seealso{
\code{\link[stats:Binomial]{rbinom()}}
}
\concept{simulate_data}
