% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-glm.R
\name{summanov}
\alias{summanov}
\title{List of Summary and Analysis of Deviance Objects for Related Univariable GLMs}
\usage{
summanov(data, .dep_var, ..., .family = binomial, .test = "Chisq")
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masked}}> quoted name of the binary dependent variable to be used as
the \acronym{LHS} of the model formula; should be numeric with values of \var{0} and \var{1}, or a two-column matrix
with the columns giving the numbers of successes and failures e.g., \code{cbind(pn, qn)}.}

\item{\dots}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> quoted name(s) of one or more factors or character vectors in
\code{.data}, to be included (or excluded)  as independent variables in the list of \acronym{GLM} analyses.}

\item{.family}{a description of the error distribution and link function to be used in the model. Can be a character
string naming a family function, a family function or the result of a call to a family function; default
\code{binomial}.}

\item{.test}{a character string, (partially) matching one of \code{"Chisq", "LRT", "Rao", "F"} or \code{"Cp"};
default \code{"Chisq"}.}
}
\value{
A \code{\link[base:list]{list}} of \code{summ_anov} objects of length equal to the number of factors or character vectors selected
using the \code{\dots} arguments. A \code{summ_anov} object is simply a list with class \code{"summ_anov"}, comprising the
following two elements: -

\item{summary}{Summary of the generalised linear model fit given by \code{\link[stats:summary.glm]{summary.glm}}.}

\item{anova}{Analysis of deviance for the generalised linear model fit given by \code{\link[stats:anova.glm]{anova.glm}}.}
}
\description{
\code{summanov()} provides a list of the summary and analysis of deviance objects for a series of related univariable
\acronym{GLMs} of data with a binary dependent variable or a (two-column) dependent variable of binomial proportions.
}
\details{
Variables in \code{.data} to be included (or excluded)  as independent variables in the list of \acronym{GLM}
analyses may be selected using the \code{\dots} argument with the<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax
of package \pkg{\link[dplyr]{dplyr}}, including use of \strong{selection helpers}.

The structure of the output list may be changed from a list of pairs into pair of lists conveniently using
\code{\link[purrr:list_transpose]{list_transpose}}. The univariable \acronym{GLMs} may then be easily compared and likewise
the univariable \acronym{GLM} anovas (analysis of deviance).

\code{\link{univ_anova}} provides a succinct summmary of the univariable analyses of deviance for all potential
categorical independent variables in \code{data}. \code{\link{anova_tbl}} also provides a succinct summmary from
the list of anovas.
}
\examples{
## Simulate Bernoulli data
(d <- list(
    iv2 = list(g = c("a", "c", "e"), h = c("b", "d", "f")),
    iv3 = list(i = c("a", "b", "c"), j = c("d", "e", "f")),
    iv4 = list(k = c("a", "b"), l = c("c", "d"), m = c("e", "f"))
) |> add_grps(bernoulli_data(levels = 6), iv, .key = _))

## Binary dependent variable
d |> summanov(dv, starts_with("iv"))
d |> summanov(dv, starts_with("iv") & !iv2)

## Binomial proportions
(d <- d |> binom_contingency(dv, starts_with("iv")))

(uva <- d |> summanov(cbind(pn, qn), num_range("iv", 2:4)))

## Change list of pairs into a pair of lists using {purrr} list_transpose()
list_transpose(uva)$summary

list_transpose(uva)$anova

rm(d, uva)
}
\seealso{
\code{\link[stats:anova.glm]{anova.glm()}}, \code{\link[purrr:list_transpose]{list_transpose()}}, \code{\link[stats:glm]{glm()}} and
\code{\link[stats:summary.glm]{summary.glm()}}; \code{\link{Print_Methods}} for S3 method for printing objects of class
\code{"summ_anov"}.

Other comp_glm: 
\code{\link{anova_tbl}()},
\code{\link{comp_glm}()},
\code{\link{univ_anova}()}
}
\concept{comp_glm}
\keyword{models}
\keyword{regression}
