% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-glm.R
\name{univ_anova}
\alias{univ_anova}
\title{Analyses of Deviance Summarising Fits of Univariable GLMs}
\usage{
univ_anova(
  data,
  .dep_var,
  .family = binomial,
  .test = c("none", "Rao", "LRT", "Chisq", "F")
)
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{.dep_var}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the binary dependent variable, which
should be \code{numeric} with values of \var{0} and \var{1}, or a two-column matrix with the columns giving the
numbers of successes and failures e.g., \code{cbind(pn, qn)}.}

\item{.family}{a family function; default \code{"binomial"}. (See \code{\link[stats:family]{family}} for details of family
functions.)}

\item{.test}{a character string, (partially) matching one of \code{"none", "Chisq", "LRT", "Rao", "F"} or \code{"Cp"};
default \code{"none"}.}
}
\value{
An object of class \verb{c("univ_anova" "announce")}, inheriting from \code{"anova"} and \code{"data.frame"} as for \code{add1()}, summarising
the differences between the fitted univariable \acronym{GLMs} and the null model.
}
\description{
\code{univ_anova()} provides a succinct summary of the analyses of deviance for all univariable GLMs based on each
possible categorical independent variable in a Bernoulli or binomial data set.
}
\details{
\code{univ_anova()} uses \code{\link[stats:add1]{add1()}} in the \pkg{\link[stats]{stats}} package to compare univariable
\acronym{GLMs} for each possible categorical independent variable (i.e., \code{"factor"} columns) in \code{.data} with
the null model. Other data types e.g. \code{"character"} vectors will be ignored and should be converted to
\code{"factor"} to be included in this analysis.
}
\examples{
d <- list(
    iv2 = list(g = c("a", "c", "e"), h = c("b", "d", "f")),
    iv3 = list(i = c("a", "b", "c"), j = c("d", "e", "f")),
    iv4 = list(k = c("a", "b"), l = c("c", "d"), m = c("e", "f"))
) |> add_grps(bernoulli_data(levels = 6), iv, .key = _) |> print()

d |> univ_anova(dv, .test = "LRT")

d |> binom_contingency(dv, starts_with("iv")) |> print() |>
     univ_anova(cbind(pn, qn), .test = "LRT")

rm(d)
}
\seealso{
\code{\link[stats:add1]{add1()}}, \code{\link[stats:anova]{anova()}}, \code{\link[stats:anova.glm]{anova.glm()}}, and
\code{\link[stats:glm]{glm()}}.

Other comp_glm: 
\code{\link{anova_tbl}()},
\code{\link{comp_glm}()},
\code{\link{summanov}()}
}
\concept{comp_glm}
\keyword{models}
\keyword{regression}
