% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-model.R
\name{var_labs}
\alias{var_labs}
\title{Format or Lookup Variable Names for Plot Titles}
\usage{
var_labs(labels)
}
\arguments{
\item{labels}{character vector containing the names of the variables to be revised.}
}
\value{
A character vector containing the revised names.
}
\description{
Vectorised labeller function used by \code{\link[=ggplot.glm_plotdata]{ggplot.glm_plotdata()}} for revising variable names
for use as subtitles in individual plots or as facet labels in faceted plots.
}
\details{
\code{var_labs} in package \pkg{\link[ParaAnita]{ParaAnita}} simply applies \code{\link[stringr:case]{str_to_title}}
to its argument. The user may override this by providing their own vectorised \code{\link[ggplot2:labeller]{labeller}}
function, see \code{labeller} under \code{\link[ggplot2:facet_wrap]{facet_wrap}} and the example.
}
\examples{

## Default labeller
c("matthew", "mark", "luke", "john") |> var_labs()

ygps <- c("year", "ygroup1", "ygroup2", "ygroup3", "ygroup4", "ygroup5", "ygroup6", "ygroup7")
mgps <- c("month", "season", "mgroup2", "mgroup3", "mgroup4", "mgroup5", "mgroup6")
demog <- c("gender", "age_group", "location", "breed")

# Vectorised function to replace terse variable names with names suitable for labelling plots
var_labs <- as_labeller(
    c(
        c("Year", paste("Year Group", seq_along(ygps[-1]))) |> set_names(ygps),
        c("Month", "Season", paste("Month Group", seq_along(mgps[-1])[-1])) |> set_names(mgps),
        c("Animal Gender", "Age Group", "Geographic Location", "Cattle Breed") |> set_names(demog)
    )
)

ygps |> var_labs()
mgps |> var_labs()
demog |> var_labs()

rm(demog, mgps, var_labs, ygps)

}
\seealso{
\code{\link[ggplot2:as_labeller]{as_labeller}}, \code{\link[ggplot2:facet_wrap]{facet_wrap}}, \code{\link[ggplot2:labeller]{labeller}}.

Other plot_model: 
\code{\link{Plot_Model}},
\code{\link{glm_plotdata}()},
\code{\link{glm_plotlist}()}
}
\concept{plot_model}
