% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{Print_Methods}
\alias{Print_Methods}
\alias{print.announce}
\alias{print.binom_contingency}
\alias{print.contingency_table}
\title{S3 Print Methods}
\usage{
\method{print}{announce}(x, ...)

\method{print}{binom_contingency}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{print}{contingency_table}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{width}{only used when \code{max.levels} is NULL, see above.}
}
\value{
The argument \code{x}.
}
\description{
S3 methods to enable printing of \code{\link{announce}}, \code{\link{binom_contingency}},
\code{\link{contingency_table}} and other objects.
}
\details{
To be completed...
These print methods return their argument \code{x} invisibly (via \code{\link[base:invisible]{invisible()}}).
}
\seealso{
\code{\link[base:print]{print()}}.

Other print: 
\code{\link{announce}()},
\code{\link{lf}()},
\code{\link{print_all}()}
}
\concept{print}
\keyword{print}
