% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{Print_Methods}
\alias{Print_Methods}
\alias{print.announce}
\alias{print.binom_contingency}
\alias{print.contingency_table}
\alias{print.odds_ratio}
\alias{print.summ_anov}
\title{S3 Print Methods}
\usage{
\method{print}{announce}(x, ...)

\method{print}{binom_contingency}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{print}{contingency_table}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{print}{odds_ratio}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{print}{summ_anov}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{width}{only used when \code{max.levels} is NULL, see above.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than the \code{print_max} \link[pillar:pillar_options]{option}.
Otherwise, will print as many rows as specified by the
\code{print_min} \link[pillar:pillar_options]{option}.}

\item{max_extra_cols}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL},
the \code{max_extra_cols} \link[pillar:pillar_options]{option} is used.
The previously defined \code{n_extra} argument is soft-deprecated.}

\item{max_footer_lines}{Maximum number of footer lines. If \code{NULL},
the \code{max_footer_lines} \link[pillar:pillar_options]{option} is used.}
}
\value{
The argument \code{x}.
}
\description{
S3 methods for printing objects of class \code{"announce"}, \code{"binom_contingency"}, \code{"contingency_table"},
\code{"odds_ratio"} and \code{"summ_anov"}.
}
\details{
These print methods return their argument \code{x} invisibly, via \code{\link[base:invisible]{invisible()}}.
}
\examples{

## print.announce() — print an 'announce' object
announce("x", lead = "Lorem ipsum dolor sit amet")

(d <- bernoulli_data())

## print.binom_contingency() — print a 'binom_contingency' object
d |> binom_contingency(dv)

(d2 <- tibble(
    iv = letters[1:4] |> sample(10, replace = TRUE) |> as.factor(),
    dv = c("Success", "Fail", "Borderline")  |> sample(10, replace = TRUE)
  ))

## print.contingency_table() — print a 'contingency_table' object
d2 |> contingency_table(dv)

## print.odds_ratio() — print an 'odds_ratio' object
d |> odds_ratio(.dep_var = dv, .ind_var = iv)

## print.summanov() — print a 'summanov' object
d |> summanov(dv, iv)

rm(d, d2)
}
\seealso{
\code{\link[base:print]{print()}}, \code{\link[tibble:formatting]{print.tbl()}}, \code{\link{binom_contingency}},
\code{\link{contingency_table}}, \code{\link{odds_ratio}} and \code{\link[=summanov]{summ_anov}}.

Other print: 
\code{\link{announce}()},
\code{\link{lf}()},
\code{\link{print_all}()}
}
\concept{print}
\keyword{print}
